+++ propagators/b3/b3_integration_test.go
@@ -44,7 +44,7 @@ func TestExtractB3(t *testing.T) {
 	}
 
 	for _, tg := range testGroup {
-		propagator := b3.B3{}
+		propagator := b3.New()
 
 		for _, tt := range tg.tests {
 			t.Run(tt.name, func(t *testing.T) {
@@ -99,7 +99,7 @@ func TestInjectB3(t *testing.T) {
 
 	for _, tg := range testGroup {
 		for _, tt := range tg.tests {
-			propagator := b3.B3{InjectEncoding: tt.encoding}
+			propagator := b3.New(b3.WithInjectEncoding(tt.encoding))
 			t.Run(tt.name, func(t *testing.T) {
 				header := http.Header{}
 				ctx := trace.ContextWithSpanContext(
@@ -130,12 +130,12 @@ func TestInjectB3(t *testing.T) {
 func TestB3Propagator_Fields(t *testing.T) {
 	tests := []struct {
 		name       string
-		propagator b3.B3
+		propagator propagation.TextMapPropagator
 		want       []string
 	}{
 		{
 			name:       "no encoding specified",
-			propagator: b3.B3{},
+			propagator: b3.New(),
 			want: []string{
 				b3TraceID,
 				b3SpanID,
@@ -145,7 +145,7 @@ func TestB3Propagator_Fields(t *testing.T) {
 		},
 		{
 			name:       "B3MultipleHeader encoding specified",
-			propagator: b3.B3{InjectEncoding: b3.B3MultipleHeader},
+			propagator: b3.New(b3.WithInjectEncoding(b3.B3MultipleHeader)),
 			want: []string{
 				b3TraceID,
 				b3SpanID,
@@ -155,14 +155,14 @@ func TestB3Propagator_Fields(t *testing.T) {
 		},
 		{
 			name:       "B3SingleHeader encoding specified",
-			propagator: b3.B3{InjectEncoding: b3.B3SingleHeader},
+			propagator: b3.New(b3.WithInjectEncoding(b3.B3SingleHeader)),
 			want: []string{
 				b3Context,
 			},
 		},
 		{
 			name:       "B3SingleHeader and B3MultipleHeader encoding specified",
-			propagator: b3.B3{InjectEncoding: b3.B3SingleHeader | b3.B3MultipleHeader},
+			propagator: b3.New(b3.WithInjectEncoding(b3.B3SingleHeader | b3.B3MultipleHeader)),
 			want: []string{
 				b3Context,
 				b3TraceID,